Migrate enumerable tests into array tests
