+++ packages/ember-runtime/tests/mixins/mutable_enumerable_test.js
@@ -1,69 +1,13 @@
-import MutableEnumerableTests from '../suites/mutable_enumerable';
 import MutableEnumerable from '../../mixins/mutable_enumerable';
-import EmberObject from '../../system/object';
-import { computed, get } from 'ember-metal';
+import ArrayProxy from '../../system/array_proxy';
+import { A } from '../../system/native_array';
 
-/*
-  Implement a basic fake mutable array.  This validates that any non-native
-  enumerable can impl this API.
-*/
-const TestMutableEnumerable = EmberObject.extend(MutableEnumerable, {
-  _content: null,
+QUnit.module('MutableEnumerable');
 
-  addObject(obj) {
-    if (this._content.indexOf(obj) >= 0) {
-      return this;
-    }
-
-    this.enumerableContentWillChange(null, [obj]);
-    this._content.push(obj);
-    this.enumerableContentDidChange(null, [obj]);
-  },
-
-  removeObject(obj) {
-    let idx = this._content.indexOf(obj);
-    if (idx < 0) {
-      return this;
-    }
-
-    this.enumerableContentWillChange([obj], null);
-    this._content.splice(idx, 1);
-    this.enumerableContentDidChange([obj], null);
-    return this;
-  },
-
-  init(ary) {
-    this._content = ary || [];
-  },
-
-  nextObject(idx) {
-    return idx >= get(this, 'length') ? undefined : this._content[idx];
-  },
-
-  length: computed(function() {
-    return this._content.length;
-  }),
-
-  slice() {
-    return this._content.slice();
-  }
+QUnit.test('should be mixed into A()', assert => {
+  assert.ok(MutableEnumerable.detect(A()));
 });
 
-
-MutableEnumerableTests.extend({
-  name: 'Basic Mutable Array',
-
-  newObject(ary) {
-    ary = ary ? ary.slice() : this.newFixture(3);
-    return new TestMutableEnumerable(ary);
-  },
-
-  // allows for testing of the basic enumerable after an internal mutation
-  mutate(obj) {
-    obj.addObject(this.getFixture(1)[0]);
-  },
-
-  toArray(obj) {
-    return obj.slice();
-  }
-}).run();
+QUnit.test('should be mixed into ArrayProxy', assert => {
+  assert.ok(MutableEnumerable.detect(ArrayProxy.create()));
+});
