+++ packages/ember-runtime/tests/suites/array/compact.js
