+++ packages/ember-runtime/tests/suites/array/every.js
