+++ packages/ember-runtime/tests/suites/array/firstObject.js
