+++ packages/ember-runtime/tests/suites/array/invoke.js
