+++ packages/ember-runtime/tests/suites/array/lastObject.js
