+++ packages/ember-runtime/tests/suites/array/reduce.js
