+++ packages/ember-runtime/tests/suites/mutable_array/removeObjects.js
